refactor(store): move recipes store to stores folder
